SystemCode/proj1Django-master/requirements.txt
appnope==0.1.0
asgiref==3.2.10
backcall==0.2.0
decorator==4.4.2
demjson==2.2.4
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
ipdb==0.13.3
ipython==7.18.1
ipython-genutils==0.2.0
jdcal==1.4.1
jedi==0.17.2
joblib==0.16.0
mysqlclient==2.0.1
numpy==1.19.2
openpyxl==3.0.5
pandas==1.1.2
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.7
ptyprocess==0.6.0
Pygments==2.7.1
python-dateutil==2.8.1
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.1.0
traitlets==5.0.4
wcwidth==0.2.5
xlrd==1.2.0
xlutils==2.0.0
xlwt==1.3.0
